Revert "Add -Wconversion to clang; fix errors found"
This reverts commit 56f3f9679267d22592376a3dae4327d68d412640.

When compiled with clang, the perft tests show no difference after the
changes in this commit.

When compiled with gcc, the changes in this commit cause a significant
slowdown in the perft tests. Given this and the noisy type-casting that
was added, I'm reverting this commit. Maybe another strategy in the
future will work better.

+++ src/Game/Clock.cpp
@@ -35,7 +35,7 @@ Game_Result Clock::punch(const Board& board) noexcept
 
     const auto time_this_punch = std::chrono::steady_clock::now();
 
-    const auto player_index = static_cast<size_t>(whose_turn);
+    const auto player_index = static_cast<int>(whose_turn);
     timers[player_index] -= (time_this_punch - time_previous_punch);
     time_previous_punch = time_this_punch;
     whose_turn = opposite(whose_turn);
@@ -73,15 +73,15 @@ Game_Result Clock::punch(const Board& board) noexcept
 
 void Clock::unpunch() noexcept
 {
-    moves_since_clock_reset[static_cast<size_t>(whose_turn)] -= 1;
-    moves_since_clock_reset[static_cast<size_t>(opposite(whose_turn))] -= 1;
+    moves_since_clock_reset[static_cast<int>(whose_turn)] -= 1;
+    moves_since_clock_reset[static_cast<int>(opposite(whose_turn))] -= 1;
     punch({});
 }
 
 void Clock::stop() noexcept
 {
     const auto time_stop = std::chrono::steady_clock::now();
-    timers[static_cast<size_t>(whose_turn)] -= (time_stop - time_previous_punch);
+    timers[static_cast<int>(whose_turn)] -= (time_stop - time_previous_punch);
     clocks_running = false;
 }
 
@@ -105,12 +105,12 @@ Clock::seconds Clock::time_left(const Piece_Color color) const noexcept
 
     if(whose_turn != color || ! clocks_running)
     {
-        return timers[static_cast<size_t>(color)];
+        return timers[static_cast<int>(color)];
     }
     else
     {
         const auto now = std::chrono::steady_clock::now();
-        return timers[static_cast<size_t>(color)] - (now - time_previous_punch);
+        return timers[static_cast<int>(color)] - (now - time_previous_punch);
     }
 }
 
@@ -118,7 +118,7 @@ size_t Clock::moves_until_reset(const Piece_Color color) const noexcept
 {
     if(move_count_reset > 0)
     {
-        return move_count_reset - moves_since_clock_reset[static_cast<size_t>(color)];
+        return move_count_reset - moves_since_clock_reset[static_cast<int>(color)];
     }
     else
     {
@@ -133,13 +133,13 @@ Piece_Color Clock::running_for() const noexcept
 
 void Clock::set_time(const Piece_Color player, const seconds new_time_seconds) noexcept
 {
-    timers[static_cast<size_t>(player)] = new_time_seconds;
+    timers[static_cast<int>(player)] = new_time_seconds;
     time_previous_punch = std::chrono::steady_clock::now();
 }
 
 void Clock::set_increment(const Piece_Color player, const seconds new_increment_time_seconds) noexcept
 {
-    increment_time[static_cast<size_t>(player)] = new_increment_time_seconds;
+    increment_time[static_cast<int>(player)] = new_increment_time_seconds;
 }
 
 void Clock::set_next_time_reset(const Piece_Color player, const size_t moves_to_reset) noexcept
@@ -150,7 +150,7 @@ void Clock::set_next_time_reset(const Piece_Color player, const size_t moves_to_
     }
     else
     {
-        move_count_reset = moves_since_clock_reset[static_cast<size_t>(player)] + moves_to_reset;
+        move_count_reset = moves_since_clock_reset[static_cast<int>(player)] + moves_to_reset;
     }
 }
 
@@ -186,7 +186,7 @@ Clock::seconds Clock::initial_time() const noexcept
 
 Clock::seconds Clock::increment(const Piece_Color color) const noexcept
 {
-    return increment_time[static_cast<size_t>(color)];
+    return increment_time[static_cast<int>(color)];
 }
 
 bool Clock::is_in_use() const noexcept
